+++ src/components/core/layout/Content/Content-types.ts
@@ -1,0 +1,4 @@
+export interface ViewProps {
+  /** - */
+  children?: React.ReactNode;
+}