+++ src/components/core/layout/index.ts
@@ -1,3 +1,4 @@
+export { default as Content } from './Content';
 export { default as Main } from './Main';
 export { default as Page } from './Page';
 export { default as Section } from './Section';