+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/), and t
   - Checkbox (inputs)
   - Form (inputs)
   - Page (layout)
+  - Content (layout)
   - Main (layout)
   - Section (layout)
   - Visibility (layout)